Settings - fix routing error & settings view added
