+++ src/components/Routes/index.tsx
@@ -1,11 +1,6 @@
-/* eslint-disable react-hooks/exhaustive-deps */
-import { useEffect } from 'react';
 import { BrowserRouter as Router, Switch } from 'react-router-dom';
-import { useMutation } from 'react-query';
 
-import { useSelector, useDispatch as useUserDispatch } from 'contexts/UserContext';
-import { actionCreators as authActions } from 'contexts/UserContext/reducer';
-import { getCurrentUserToken, getUser, setApiTokenHeader } from 'services/AuthService';
+import { useSelector } from 'contexts/UserContext';
 
 import Suspense from '../Suspense';
 
@@ -14,29 +9,6 @@ import RouteItem from './components/RouteItem';
 import styles from './styles.module.scss';
 
 function Routes() {
-  const userDispatch = useUserDispatch();
-
-  const { mutate } = useMutation(() => getUser(), {
-    onSettled: (data) => {
-      if (data?.data && data?.ok) {
-        const {
-          data: { user }
-        } = data;
-        if (user) {
-          userDispatch(authActions.setUser(user));
-        }
-      }
-    }
-  });
-
-  useEffect(() => {
-    const token = getCurrentUserToken();
-    if (token) {
-      setApiTokenHeader(token);
-      mutate();
-    }
-  }, []);
-
   const user = useSelector((state) => state.user);
 
   return (
