+++ src/screens/Settings/index.tsx
@@ -1,0 +1,11 @@
+import Layout from 'components/Layout';
+
+function Settings() {
+  return (
+    <Layout>
+      <h1>User settings</h1>
+    </Layout>
+  );
+}
+
+export default Settings;
